Pulled a lot of the action handling out into a separate class. This has made the ApplicationController much shorter, but the ActionHandler is a bit big now :-)
